+++ src-docs/src/views/flex/direction.js
@@ -1,0 +1,14 @@
+import React from 'react';
+
+import {
+  EuiFlexGroup,
+  EuiFlexItem,
+} from '../../../../src/components';
+
+export default () => (
+  <EuiFlexGroup direction="column">
+    <EuiFlexItem>Content grid item</EuiFlexItem>
+    <EuiFlexItem>Another content grid item</EuiFlexItem>
+    <EuiFlexItem>Using the column direction</EuiFlexItem>
+  </EuiFlexGroup>
+);