+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 ## [`master`](https://github.com/elastic/eui/tree/master)
 
 - Added `EuiBetaBadge` for non-GA labelling including options to add it to `EuiCard` and `EuiKeyPadMenuItem` ([#705](https://github.com/elastic/eui/pull/705))
-- Added `EuiDescriptiveFormRow` component, a wrapper around `EuiFormRow` to display it horizontally on desktop ([#707](https://github.com/elastic/eui/pull/707))
+- Added `direction` prop to EuiFlexGroup ([#711](https://github.com/elastic/eui/pull/711))
+- Added `EuiEmptyPrompt` which can be used as a placeholder over empty tables and lists ([#711](https://github.com/elastic/eui/pull/711))
+- Added `EuiTabbedContent` ([#737](https://github.com/elastic/eui/pull/737))
+- Added `EuiDescriptiveFormRow` component, a wrapper around `EuiFormRow`(s) to display it horizontally on desktop ([#707](https://github.com/elastic/eui/pull/707))
+
+**Bug fixes**
+
+- Fixed `EuiTableRowCell` from overwriting its child element's `className` [#709](https://github.com/elastic/eui/pull/709)
+- Allow `EuiContextMenuPanel`s to update when their `children` changes ([#710](https://github.com/elastic/eui/pull/710))
 
 ## [`0.0.44`](https://github.com/elastic/eui/tree/v0.0.44)